move: admin component to root level;
