+++ src/app/admin/admin-categories/admin-categories.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { CategoryService } from "../../../services/category.service";
+import { CategoryService } from "../../services/category.service";
 import { CategoryFormService } from "./category-form/category-form.service";
 
 @Component({
