+++ src/app/admin/admin-categories/category-form/category-form.service.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import {Category} from "../../../../classes/category";
+import { Category } from "../../../classes/category";
 
 @Injectable()
 export class CategoryFormService {
