+++ src/app/app.module.ts
@@ -17,13 +17,13 @@ import { HeaderComponent } from './components/header/header.component';
 import { DashboardComponent } from './components/dashboard/dashboard.component';
 import { RegisterComponent } from './components/register/register.component';
 import { AccountActivationComponent } from './components/account-activation/account-activation.component';
-import { AdminComponent } from './components/admin/admin.component';
+import { AdminComponent } from './admin/admin.component';
 import { AdminGuard } from "./core/admin.guard";
-import { AdminCategoriesComponent } from './components/admin/admin-categories/admin-categories.component';
-import { AdminTasksComponent } from './components/admin/admin-tasks/admin-tasks.component';
+import { AdminCategoriesComponent } from './admin/admin-categories/admin-categories.component';
+import { AdminTasksComponent } from './admin/admin-tasks/admin-tasks.component';
 import { CategoryService } from "./services/category.service";
-import { CategoryFormComponent } from './components/admin/admin-categories/category-form/category-form.component';
-import { CategoryFormService } from './components/admin/admin-categories/category-form/category-form.service';
+import { CategoryFormComponent } from './admin/admin-categories/category-form/category-form.component';
+import { CategoryFormService } from './admin/admin-categories/category-form/category-form.service';
 import { TaskService } from './services/task.service';
 import { AssessmentTypeService } from './services/assessment-type.service';
